% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlm.R
\name{mlm}
\alias{mlm}
\title{Fit an OLS multiple regression model}
\usage{
mlm(data, response_col_name)
}
\arguments{
\item{data}{A data frame with only numeric variables and no missing data.}

\item{response_col_name}{The name of the response variable.}
}
\value{
An object of class "mlm".
}
\description{
Fit an OLS multiple regression model
}
