% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlm.R
\name{print.mlm}
\alias{print.mlm}
\title{Print basic information about an mlm object}
\usage{
\method{print}{mlm}(mlm_mod, ...)
}
\arguments{
\item{mlm_mod}{An object of class "mlm"}

\item{...}{}
}
\description{
Print basic information about an mlm object
}
