% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlm.R
\name{predict.mlm}
\alias{predict.mlm}
\title{Generate predictions from an mlm model}
\usage{
\method{predict}{mlm}(object, newdata, ...)
}
\arguments{
\item{object}{An object of class "mlm"}

\item{...}{}
}
\description{
Generate predictions from an mlm model
}
